Refactor: Consolidate emotion analysis components
Integrates Korean text and EEG emotion processing into a unified engine. Introduces FFT-based bandpower calculation for EEG analysis and refactors existing components for better modularity and testing.
